dashboard/requirements.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
google-generativeai==0.6.0
python-dotenv==1.0.1
numpy==1.26.4